+++ src/types/navigation.ts
@@ -22,7 +22,9 @@ export interface ModalPropsMap {
 	Alert: AlertModalProps
 	EditLyrics: { uniqueKey: string; lyrics: ParsedLrc }
 	SleepTimer: undefined
+	SaveQueueToPlaylist: { trackIds: string[] }
 	DonationQR: { type: 'wechat' }
+	PlaybackSpeed: undefined
 }
 
 export type ModalKey = keyof ModalPropsMap